Some refactoring of the calendar tests
Due to the way these were implemented, they would fail when the test were
run on some days of the month since days in the future are not clickable.
This just reduces the number of days when the tests will fail to the 1st
of every month, since faking the current date seems to break the tests and
they never complete.