Better handling of dynamic translations
The names of objects like muscles, equipment, etc. which come from the database /
the server are translated. However, if the manual process is not run (e.g. because
this is a local instance and the administrator does not commit the changes, etc.)
we would show the internal translation key in the UI, instead of the original name.